streaming-scripts/csv-generator/requirements.txt
fake-factory==0.7.2
numpy==1.11.2
Faker==0.7.3
pytz==2016.7
tzlocal==1.3.0
google-cloud-storage==1.14.0